# Reduced number of allocations for dense matrix exp