This PR reduces the number of allocations by the dense matrix exponential `exp(A)` by about half for the case of `opnorm(A,1)` > 2.1. It also removes two allocations for the case of `opnorm(A,1)` <= 2.1.

There are still a few allocations that could be removed, i.e., by replacing
`U = U*A` with `U, tmp1 = mul!(tmp1, A, U), A`, not sure about the weighting in the readability/performance trade-off? 


```julia
for n in [10,30,100,300]
    A = randn(n,n)
    A = 6A/opnorm(A,1)
    @btime exp($A)
end
```
Before:
```
  15.159 μs (17 allocations: 12.70 KiB)
  82.787 μs (17 allocations: 101.39 KiB)
  572.812 μs (31 allocations: 1.07 MiB)
  9.999 ms (31 allocations: 9.62 MiB)
```
After:
```
  14.063 μs (10 allocations: 6.58 KiB)
  77.001 μs (10 allocations: 51.08 KiB)
  537.138 μs (17 allocations: 549.12 KiB)
  8.962 ms (17 allocations: 4.81 MiB)
```